further split snapshot related behaviours.
